requirements.txt
litellm>=1.66.1
crawl4ai>=0.5.0
google-adk>=0.1.0
pytest>=8.3.5
pandas>=2.2.3
tabulate>=0.9.0
scikit-learn1.6.1